% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_validateFunc_headers.R
\name{A15_validateFunction}
\alias{A15_validateFunction}
\title{A15 validation function}
\usage{
A15_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
A15 validation function
}
