% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_data.R
\docType{data}
\name{A1_formats}
\alias{A1_formats}
\title{Table A1. BSRN station-to-archive file format.}
\format{
An object of class \code{data.frame} with 164 rows and 8 columns.
}
\usage{
A1_formats
}
\description{
All logical records are compulsory definitions. The file is identified by the
station id no., the year and the month in logical record 0001. The dates of
change in logical records 0002, 0004, 0005, 0006, 0007, 0008, and 0009 are given
by day, hour, and minute with ranges 1... 31, 0... 23, and 0... 59. The dates of
measurement in logical records 0100, 0200, ... are given by day and minute with
ranges 1... 31 and 0... 1439 also for quantities measured in hour intervals
}
\keyword{datasets}
