% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{A2_bsrnStations}
\alias{A2_bsrnStations}
\title{Table A2. BSRN stations.}
\format{
An object of class \code{data.frame} with 76 rows and 10 columns.
}
\usage{
A2_bsrnStations
}
\description{
For more information see: http://www.bsrn.awi.de/en/stations/listings/.
}
\keyword{datasets}
