% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_validateFunc_headers.R
\name{A30_validateFunction}
\alias{A30_validateFunction}
\title{A30 validation function}
\usage{
A30_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
A30 validation function
}
