% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_data.R
\docType{data}
\name{A3_quantities}
\alias{A3_quantities}
\title{Table A3. Quantity measured.}
\format{
An object of class \code{numeric} of length 28.
}
\usage{
A3_quantities
}
\description{
Every radiation value is measured by exactly one radiation instrument.
If a value in height is missing, the quantity is measured only once at
standard height. The id. no. of instruments not measured at standard height
consists of the id. no. measured at standard height followed by 6 numericals
expressing the height of the instruments above ground in cm.
}
\keyword{datasets}
