% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{A4_surfaces}
\alias{A4_surfaces}
\title{Table A4. Types of surface.}
\format{
An object of class \code{numeric} of length 21.
}
\usage{
A4_surfaces
}
\description{
Table A4. Types of surface.
}
\keyword{datasets}
