% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{A5_topographies}
\alias{A5_topographies}
\title{Table A5. Types of topography.}
\format{
An object of class \code{numeric} of length 8.
}
\usage{
A5_topographies
}
\description{
Table A5. Types of topography.
}
\keyword{datasets}
