% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_data.R
\docType{data}
\name{A6_pyrgeometers}
\alias{A6_pyrgeometers}
\title{Table A6. Pyrgeometer body temperature compensation codes.}
\format{
An object of class \code{numeric} of length 4.
}
\usage{
A6_pyrgeometers
}
\description{
Table A6. Pyrgeometer body temperature compensation codes.
}
\keyword{datasets}
