% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{A7_pyrgeometers}
\alias{A7_pyrgeometers}
\title{Table A7. Pyrgeometer dome temperature compensation codes.}
\format{
An object of class \code{numeric} of length 8.
}
\usage{
A7_pyrgeometers
}
\description{
Table A7. Pyrgeometer dome temperature compensation codes.
}
\keyword{datasets}
