% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_validateFunc_headers.R
\name{A80_validateFunction}
\alias{A80_validateFunction}
\title{A80 validation function}
\usage{
A80_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
A80 validation function
}
