% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-validateFunctions.R
\name{A_validateFunction}
\alias{A_validateFunction}
\title{Character validation function}
\usage{
A_validateFunction(value, maxLenght = Inf)
}
\arguments{
\item{value}{value to validate}

\item{maxLenght}{maximun lenght of value}
}
\description{
Character validation function
}
