% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utils.R
\name{applyFormat}
\alias{applyFormat}
\title{Generic function to apply the right format on the value}
\usage{
applyFormat()
}
\arguments{
\item{varName}{char - Variable Name}
}
\description{
Internal function.
}
