% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-validateFunctions.R
\name{azimuth_validateFunction}
\alias{azimuth_validateFunction}
\title{azimuth validation function}
\usage{
azimuth_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
azimuth validation function
}
