% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_validateFunc_headers.R
\name{C_validateFunction}
\alias{C_validateFunction}
\title{Numerical validation function}
\usage{
C_validateFunction(value)
}
\arguments{
\item{value}{Value to validate}
}
\value{
NULL if validation passes, otherwise an error message
}
\description{
Numerical validation function
}
