% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-validateFunctions.R
\name{date_validateFunction}
\alias{date_validateFunction}
\title{date validation function}
\usage{
date_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
format : (MM/DD/YY)
}
