% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_validateFunc_headers.R
\name{dome_validateFunction}
\alias{dome_validateFunction}
\title{pyrgeometer dome validation function}
\usage{
dome_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
pyrgeometer dome validation function
}
