% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utils.R
\name{dot}
\alias{dot}
\title{Put a dot on a variable name}
\usage{
dot(varName)
}
\arguments{
\item{varName}{char - Variable Name}
}
\value{
The variable name with a dot
}
\description{
This function put a dot behind the variable name : "foo" => ".foo".
}
\examples{
dot('foo')
}
