% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-validateFunctions.R
\name{elevation_validateFunction}
\alias{elevation_validateFunction}
\title{elevation validation function}
\usage{
elevation_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
elevation validation function
}
