% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_validateFunc_headers.R
\name{email_validateFunction}
\alias{email_validateFunction}
\title{email validation function}
\usage{
email_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
email validation function
}
