% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_validateFunc_headers.R
\name{F12.4_validateFunction}
\alias{F12.4_validateFunction}
\title{F12.4 validation function}
\usage{
F12.4_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
F12.4 validation function
}
