% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-validateFunctions.R
\name{F7.3_validateFunction}
\alias{F7.3_validateFunction}
\title{F7.3 validation function}
\usage{
F7.3_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
F7.3 validation function
}
