% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_validateFunc_headers.R
\name{F_validateFunction}
\alias{F_validateFunction}
\title{Fixed point format validation fonction}
\usage{
F_validateFunction(value, w, d)
}
\arguments{
\item{value}{value to validate}

\item{w}{w of Fw.d}

\item{d}{d of Fw.d}
}
\description{
Fixed point format validation fonction
}
