% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utils.R
\name{genericGetFormatedValue}
\alias{genericGetFormatedValue}
\title{Generic function to get formated value}
\usage{
genericGetFormatedValue(varName)
}
\arguments{
\item{varName}{char - Variable Name}
}
\description{
Internal function.
}
