% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utils.R
\name{genericGetMissingCode}
\alias{genericGetMissingCode}
\title{Generic function to get missing code}
\usage{
genericGetMissingCode(varName)
}
\arguments{
\item{varName}{char - Variable Name}
}
\description{
Internal function.
}
