% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-utils.R
\name{genericIsMandatory}
\alias{genericIsMandatory}
\title{Generic function to control mandatory}
\usage{
genericIsMandatory(varName)
}
\arguments{
\item{varName}{char - Variable Name}
}
\description{
Internal function.
Allow to create an R6 public method to know if private$varName$mandatory.
}
