% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utils.R
\name{genericMandatories}
\alias{genericMandatories}
\title{Generic function to list all mandatory variable}
\usage{
genericMandatories()
}
\description{
Internal function.
This function give a list of all mandatory variable names.
}
