% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utils.R
\name{genericMissings}
\alias{genericMissings}
\title{Generic function to list all missing vairable}
\usage{
genericMissings()
}
\description{
Internal function.
This function give a list of all missing variable names, i.e. the variable
is empty and mandatory.
}
