% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utils.R
\name{genericSetDefault}
\alias{genericSetDefault}
\title{Generic function to set value to it default value}
\usage{
genericSetDefault(varName)
}
\arguments{
\item{varName}{char - Variable Name}
}
\description{
Internal function.
}
