% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_utils.R
\name{genericValidateFunction}
\alias{genericValidateFunction}
\title{Generic function to validate data}
\usage{
genericValidateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
Internal function.
}
