% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utils.R
\name{genericValueActiveBinding}
\alias{genericValueActiveBinding}
\title{Generic active binding for value}
\usage{
genericValueActiveBinding(varName)
}
\arguments{
\item{varName}{char - Variable Name}
}
\value{
Binding for the varName.
}
\description{
Internal function.
Create binding between private variable and R6 object.
}
\examples{
\dontrun{
username = genericValueActiveBinding(".username")
}
}
