% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-utils.R
\name{getActiveBindings}
\alias{getActiveBindings}
\title{Return actives bindings for a specific logical record}
\usage{
getActiveBindings(lr)
}
\arguments{
\item{lr}{char logical record ("LR0001", "LR0002", ...)}
}
\value{
list of actives bindings
}
\description{
Return actives bindings for a specific logical record
}
