% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-utils.R
\name{getAzimuthElevation}
\alias{getAzimuthElevation}
\title{Get the BSRN format for LR0004 azimuth and elevation}
\usage{
getAzimuthElevation(azimuth = NULL, elevation = NULL)
}
\arguments{
\item{azimuth}{(string) azimuth at format "A1,A2,...,An"}

\item{elevation}{(string) elevation at format "E1,E2,...,En"}
}
\value{
(string) Azimuth and elevation at LR0004 format
}
\description{
Internal function.
}
