% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_utils.R
\name{getFormatValue}
\alias{getFormatValue}
\title{Generic function to apply the right format on the value}
\usage{
getFormatValue(varName)
}
\arguments{
\item{varName}{char - Variable Name}
}
\description{
Internal function.
}
