% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_utils.R
\name{getParams}
\alias{getParams}
\title{Return parameters for a specific logical record}
\usage{
getParams(lr)
}
\arguments{
\item{lr}{char logical record ("LR0001", "LR0002", ...)}
}
\value{
params list
}
\description{
Return parameters for a specific logical record
}
\examples{
\dontrun{
getParams("LR0001")
}

}
