% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-utils.R
\name{getPrivateVars}
\alias{getPrivateVars}
\title{Return privates variables for a specific logical record}
\usage{
getPrivateVars(lr)
}
\arguments{
\item{lr}{char logical record ("LR0001", "LR0002", ...)}
}
\value{
list of private variables
}
\description{
Return privates variables for a specific logical record
}
