% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-utils.R
\name{getPublicMethods}
\alias{getPublicMethods}
\title{Return public methods for a specific logical record}
\usage{
getPublicMethods(lr)
}
\arguments{
\item{lr}{char logical record ("LR0001", "LR0002", ...)}
}
\value{
list of private variables
}
\description{
Return public methods for a specific logical record
}
