% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_validateFunc_headers.R
\name{hour_validateFunction}
\alias{hour_validateFunction}
\title{hour validation function}
\usage{
hour_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
hour validation function
}
