% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_validateFunc_headers.R
\name{I3_validateFunction}
\alias{I3_validateFunction}
\title{I3 validation function}
\usage{
I3_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
I3 validation function
}
