% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_validateFunc_headers.R
\name{I4_validateFunction}
\alias{I4_validateFunction}
\title{I4 validation function}
\usage{
I4_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
I4 validation function
}
