% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_validateFunc_headers.R
\name{I5_validateFunction}
\alias{I5_validateFunction}
\title{I5 validation function}
\usage{
I5_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
I5 validation function
}
