% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_validateFunc_headers.R
\name{I8_validateFunction}
\alias{I8_validateFunction}
\title{I8 validation function}
\usage{
I8_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
I8 validation function
}
