% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-validateFunctions.R
\name{I_validateFunction}
\alias{I_validateFunction}
\title{Integer validation function}
\usage{
I_validateFunction(value, digits, min = 0, max = (10^digits) - 1)
}
\arguments{
\item{value}{value to validate}

\item{digits}{number of digits}
}
\description{
Integer validation function
}
