% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_validateFunc_headers.R
\name{L_validateFunction}
\alias{L_validateFunction}
\title{Logical validation function}
\usage{
L_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
Logical validation function
}
