% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_validateFunc_headers.R
\name{latitude_validateFunction}
\alias{latitude_validateFunction}
\title{latitude validation function}
\usage{
latitude_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
latitude validation function
}
