% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-LR0001.R
\name{LR0001}
\alias{LR0001}
\title{R6 Class representing a logical record 0001}
\description{
R6 Class representing a logical record 0001

R6 Class representing a logical record 0001
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-isMandatory}{\code{LR0001$isMandatory()}}
\item \href{#method-isMissing}{\code{LR0001$isMissing()}}
\item \href{#method-mandatories}{\code{LR0001$mandatories()}}
\item \href{#method-missings}{\code{LR0001$missings()}}
\item \href{#method-isValuesMissing}{\code{LR0001$isValuesMissing()}}
\item \href{#method-setDefault}{\code{LR0001$setDefault()}}
\item \href{#method-getMissingCode}{\code{LR0001$getMissingCode()}}
\item \href{#method-getLabel}{\code{LR0001$getLabel()}}
\item \href{#method-showBsrnFormat}{\code{LR0001$showBsrnFormat()}}
\item \href{#method-print}{\code{LR0001$print()}}
\item \href{#method-getFormatValue}{\code{LR0001$getFormatValue()}}
\item \href{#method-getBsrnFormat}{\code{LR0001$getBsrnFormat()}}
\item \href{#method-new}{\code{LR0001$new()}}
\item \href{#method-clone}{\code{LR0001$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isMandatory"></a>}}
\if{latex}{\out{\hypertarget{method-isMandatory}{}}}
\subsection{Method \code{isMandatory()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LR0001$isMandatory(var)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isMissing"></a>}}
\if{latex}{\out{\hypertarget{method-isMissing}{}}}
\subsection{Method \code{isMissing()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LR0001$isMissing(var)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mandatories"></a>}}
\if{latex}{\out{\hypertarget{method-mandatories}{}}}
\subsection{Method \code{mandatories()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LR0001$mandatories()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-missings"></a>}}
\if{latex}{\out{\hypertarget{method-missings}{}}}
\subsection{Method \code{missings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LR0001$missings()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isValuesMissing"></a>}}
\if{latex}{\out{\hypertarget{method-isValuesMissing}{}}}
\subsection{Method \code{isValuesMissing()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LR0001$isValuesMissing()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setDefault"></a>}}
\if{latex}{\out{\hypertarget{method-setDefault}{}}}
\subsection{Method \code{setDefault()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LR0001$setDefault(var)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMissingCode"></a>}}
\if{latex}{\out{\hypertarget{method-getMissingCode}{}}}
\subsection{Method \code{getMissingCode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LR0001$getMissingCode(var)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getLabel"></a>}}
\if{latex}{\out{\hypertarget{method-getLabel}{}}}
\subsection{Method \code{getLabel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LR0001$getLabel(var)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-showBsrnFormat"></a>}}
\if{latex}{\out{\hypertarget{method-showBsrnFormat}{}}}
\subsection{Method \code{showBsrnFormat()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LR0001$showBsrnFormat(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LR0001$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFormatValue"></a>}}
\if{latex}{\out{\hypertarget{method-getFormatValue}{}}}
\subsection{Method \code{getFormatValue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LR0001$getFormatValue(var)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getBsrnFormat"></a>}}
\if{latex}{\out{\hypertarget{method-getBsrnFormat}{}}}
\subsection{Method \code{getBsrnFormat()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LR0001$getBsrnFormat(listSensor = c("2", "3", "4", "5", "21", "22", "23"))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LR0001$new(stationNumber = NULL, month = NULL, year = NULL, version = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LR0001$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
