% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-LR0001.R
\name{lr0001GetBsrnFormat}
\alias{lr0001GetBsrnFormat}
\title{LR0001 get BSRN format function}
\usage{
lr0001GetBsrnFormat(listSensor = c("2", "3", "4", "5", "21", "22", "23"))
}
\arguments{
\item{listSensor}{List of sensor use in the file}
}
\value{
A char with the BSRN format
}
\description{
LR0001 get BSRN format function
}
