% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_R6Class_headers.R
\name{LR0002}
\alias{LR0002}
\title{R6 Class representing BSRN LR0002 (Logical Record 0002)}
\arguments{
\item{scientistChange}{Is scientist changed this month ?}

\item{scientistChangeDay}{Day}

\item{scientistChangeHour}{Hour}

\item{scientistChangeMinute}{Minute}

\item{scientistName}{Scientist name *}

\item{scientistTel}{Scientist telephone no. *}

\item{scientistFax}{Scientist fax no. *}

\item{scientistTcpip}{Scientist TCP/IP no.}

\item{scientistMail}{Scientist e-mail address}

\item{scientistAddress}{Scientist address}

\item{deputyChange}{Is deputy changed this month ?}

\item{deputyChangeDay}{Day}

\item{deputyChangeHour}{Hour}

\item{deputyChangeMinute}{Minute}

\item{deputyName}{Deputy name}

\item{deputyTel}{Deputy telephone no.}

\item{deputyFax}{Deputy fax no.}

\item{deputyTcpip}{Deputy TCP/IP no.}

\item{deputyMail}{Deputy e-mail address}

\item{deputyAddress}{Deputy address}
}
\description{
A LR0002 has deputy and scientist informations.
\itemize{
  \item LR0002$new() : Create a new LR0002 object
  \item isMandatory() : Is variable version is mandatory ?
  \item isMissing() : Is variable version is missing ?
  \item isValuesMissing() : Is any value is missing ?
  \item mandatories() : List of mandatory variable.
  \item missings() : List of missing variables.
  \item setDefault() : Set a variable to default value.
  \item getMissingCode() : Get the missing code for the variable.
  \item getLabel() : Get the label for the variable.
  \item getFormatValue () : Get value with the correct format.
  \item getBsrnFormat() : Get the BSRN format of the object.
  \item showBsrnFormat() : Print the BSRN format.
  \item print() : Generic print method.
}
}
\examples{
lr0002 = LR0002$new(scientistChange = FALSE, scientistChangeDay = -1, scientistChangeHour = -1, scientistChangeMinute = -1, scientistName = NULL, scientistTel = NULL, scientistFax = NULL, scientistTcpip = NULL, scientistMail = NULL, scientistAddress = NULL, deputyChange = FALSE, deputyChangeDay = -1, deputyChangeHour = -1, deputyChangeMinute = -1, deputyName = NULL, deputyTel = NULL, deputyFax = NULL, deputyTcpip = NULL, deputyMail = NULL, deputyAddress = NULL)
lr0002$scientistChange = TRUE
lr0002$stationNumber
lr0002$mandatories()
lr0002$missings()
lr0002$isValuesMissing()
lr0002$setDefault("month")
lr0002$month
lr0002$month = 1
lr0002$getMissingCode("month")
lr0002$getLabel("month")
lr0002$getBsrnFormat()
lr0002$showBsrnFormat()
lr0002

}
