% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-bsrnFormats.R
\name{lr0002GetBsrnFormat}
\alias{lr0002GetBsrnFormat}
\title{LR0002 get BSRN format function}
\usage{
lr0002GetBsrnFormat()
}
\value{
A char with the BSRN format
}
\description{
LR0002 get BSRN format function
}
