% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_R6Class_headers.R
\name{LR0003}
\alias{LR0003}
\title{R6 Class representing BSRN LR0003 (Logical Record 0003)}
\arguments{
\item{message}{Messages not to be inserted in the BSRN database}
}
\description{
A LR0003 contain a message.
}
\details{
\itemize{
  \item LR0003$new() : Create a new LR0003 object
  \item isMandatory() : Is variable version is mandatory ?
  \item isMissing() : Is variable version is missing ?
  \item isValuesMissing() : Is any value is missing ?
  \item mandatories() : List of mandatory variable.
  \item missings() : List of missing variables.
  \item setDefault() : Set a variable to default value.
  \item getFormatValue () : Get value with the correct format.
  \item getBsrnFormat() : Get the BSRN format of the object.
  \item showBsrnFormat() : Print the BSRN format.
  \item print() : Generic print method.
}
}
\examples{
lr0003 <- LR0003$new(message = "Something important")

lr0003$getBsrnFormat()
lr0003$showBsrnFormat()
lr0003
}
