% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-LR0003.R
\name{lr0003GetBsrnFormat}
\alias{lr0003GetBsrnFormat}
\title{LR0003 get BSRN format function}
\usage{
lr0003GetBsrnFormat()
}
\value{
A char with the BSRN format
}
\description{
LR0003 get BSRN format function
}
