% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_bsrnFormats_headers.R
\name{lr0004GetBsrnFormat}
\alias{lr0004GetBsrnFormat}
\title{LR0004 get BSRN format function}
\usage{
lr0004GetBsrnFormat()
}
\value{
A char with the BSRN format
}
\description{
LR0004 get BSRN format function
}
