% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-bsrnFormats.R
\name{lr0005GetBsrnFormat}
\alias{lr0005GetBsrnFormat}
\title{LR0005 get BSRN format function}
\usage{
lr0005GetBsrnFormat()
}
\value{
A char with the BSRN format
}
\description{
LR0005 get BSRN format function
}
