% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-bsrnFormats.R
\name{lr0006GetBsrnFormat}
\alias{lr0006GetBsrnFormat}
\title{LR0006 get BSRN format function}
\usage{
lr0006GetBsrnFormat()
}
\value{
A char with the BSRN format
}
\description{
LR0006 get BSRN format function
}
