% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-bsrnFormats.R
\name{lr0007GetBsrnFormat}
\alias{lr0007GetBsrnFormat}
\title{LR0007 get BSRN format function}
\usage{
lr0007GetBsrnFormat(synop)
}
\value{
A char with the BSRN format
}
\description{
LR0007 get BSRN format function
}
