% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_bsrnFormats_headers.R
\name{lr0008GetBsrnFormat}
\alias{lr0008GetBsrnFormat}
\title{LR0008 get BSRN format function}
\usage{
lr0008GetBsrnFormat(anyChange = FALSE, printLr = FALSE, LR0009Format = FALSE)
}
\arguments{
\item{anyChange}{boolean if LR0008 changed}

\item{printLr}{boolean to print first row of LR}

\item{LR0009Format}{boolean display LR0009 ?}
}
\value{
A char with the BSRN format
}
\description{
LR0008 get BSRN format function
}
