% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-bsrnFormats.R
\name{lr0009GetBsrnFormat}
\alias{lr0009GetBsrnFormat}
\title{LR0009 get BSRN format function}
\usage{
lr0009GetBsrnFormat(change = FALSE, printLr = FALSE)
}
\value{
A char with the BSRN format
}
\description{
LR0009 get BSRN format function
}
