% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_validateFunc_datas.R
\name{LR0100_validateFunction}
\alias{LR0100_validateFunction}
\title{LR0100 validation function}
\usage{
LR0100_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
LR0100 validation function
}
