% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_validateFunc_datas.R
\name{LR4000_validateFunction}
\alias{LR4000_validateFunction}
\title{LR4000 validation function}
\usage{
LR4000_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
LR4000 validation function
}
