% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_R6Class_headers.R
\name{LR4000CONST}
\alias{LR4000CONST}
\title{R6 Class representing BSRN LR4000CONST (Logical Record 4000 constant)}
\arguments{
\item{serialNumber_Manufacturer}{serial number appears in the certificate plate}

\item{serialNumber_WRMC}{the one used in your station-to-archive files}

\item{certificateCodeID}{Certificate Code ID (type 1 or 2)}

\item{yyyymmdd}{date of the calibration certificate issued}

\item{manufact}{KZ (Kipp and Zonen), EP (Eppley), HF (Hukseflux), ...}

\item{model}{CH1,CH1P,CM11,CM21,CM21P,CM22,CM22P,CG4,CGR4, ...}

\item{C}{thermopile responsivity}

\item{k0}{general equation of the pyrgeometer}

\item{k1}{general equation of the pyrgeometer}

\item{k2}{general equation of the pyrgeometer}

\item{k3}{general equation of the pyrgeometer}

\item{f}{general equation of the pyrgeometer}
}
\description{
A LR4000CONST contain object.
}
\examples{
lr4000 <- LR4000$new()

lr4000$yearMonth <- "2012-01"
lr4000$bodyT_down <- rep(1, 44640)

lr4000
lr4000$showBsrnFormat(changed = TRUE)
}
