% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_bsrnFormats_headers.R
\name{lr4000constGetBsrnFormat}
\alias{lr4000constGetBsrnFormat}
\title{LR4000CONST get BSRN format function}
\usage{
lr4000constGetBsrnFormat(method = 1)
}
\arguments{
\item{method}{A character or numeric value indicating which method to use for compiling the `@LR4000CONST` metadata:
- If `1`, the function will use **Method 1**, where all calibration constants and metadata are manually inputted directly in the metadata line (e.g., `@LR4000CONST, s/n, WMO, CertificateCodeID, C, k0, k1, k2, k3, f`).
- If `2`, the function will use **Method 2**, where the calibration certificate is stored centrally, and the metadata line includes a reference to the centrally stored calibration document (e.g., `@LR4000CONST, s/n, WMO, CAL_date_manufacturer_model_serial_WMO,& C, k0, k1, k2, k3, f`).
Method 2 is preferred if calibration documents are stored in a centralized archive.}
}
\value{
A char with the BSRN format
}
\description{
LR4000CONST get BSRN format function
}
