% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_bsrnFormats_datas.R
\name{lr4000GetBsrnFormat}
\alias{lr4000GetBsrnFormat}
\title{LR4000 get BSRN format function}
\usage{
lr4000GetBsrnFormat(changed)
}
\arguments{
\item{changed}{boolean did LR4000 changed this month ?}
}
\value{
A char with the BSRN format
}
\description{
LR4000 get BSRN format function
}
