% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_validateFunc_headers.R
\name{minute_validateFunction}
\alias{minute_validateFunction}
\title{minute validation function}
\usage{
minute_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
minute validation function
}
