% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_utils.R
\name{numberOfDays}
\alias{numberOfDays}
\title{Number of days into a month}
\usage{
numberOfDays(yearMonth)
}
\arguments{
\item{yearMonth}{(string) year month format "YYYY-MM". Ex : "2012-02"}
}
\value{
(numeric) number of day into the month
}
\description{
Number of days into a month
}
