% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_utils.R
\name{numberOfMinutes}
\alias{numberOfMinutes}
\title{Number of minutes into a month}
\usage{
numberOfMinutes(yearMonth)
}
\arguments{
\item{yearMonth}{(string) year month format "YYYY-MM". Ex : "2012-02"}
}
\value{
(numeric) number of minutes into the month
}
\description{
Number of minutes into a month
}
