% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-utils.R
\name{parseEval}
\alias{parseEval}
\title{Parse text containing R code and eval it.}
\usage{
parseEval(text)
}
\arguments{
\item{text}{text containing R code}
}
\value{
evaluation of R code
}
\description{
Parse text containing R code and eval it.
}
