% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-validateFunctions.R
\name{quantities_validateFunction}
\alias{quantities_validateFunction}
\title{radiation quantities measured validation function}
\usage{
quantities_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
radiation quantities measured validation function
}
