% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_utils.R
\name{r__ActiveBinding}
\alias{r__ActiveBinding}
\title{Generic read only active binding}
\usage{
r__ActiveBinding(varName)
}
\arguments{
\item{varName}{char - Variable Name}
}
\value{
Binding for the varName.
}
\description{
Internal function.
Create read and write binding between private variable and R6 object.
}
\examples{
\dontrun{
username = r__ActiveBinding("username")
}

}
