% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_utils.R
\name{rw_ActiveBinding}
\alias{rw_ActiveBinding}
\title{Generic read and write active binding}
\usage{
rw_ActiveBinding(varName)
}
\arguments{
\item{varName}{char - Variable Name}
}
\value{
Binding for the varName.
}
\description{
Internal function.
Create read and write binding between private variable and R6 object.
}
\examples{
\dontrun{
username = rw_ActiveBinding("username")
}

}
