% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utils.R
\name{stopIfValuesMissing}
\alias{stopIfValuesMissing}
\title{Stop is some mandatory variables are missing}
\usage{
stopIfValuesMissing(message = NULL, self)
}
\arguments{
\item{message}{A message if stop}

\item{self}{self of the R6 object}
}
\description{
Stop is some mandatory variables are missing
}
