% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-validateFunctions.R
\name{surface_validateFunction}
\alias{surface_validateFunction}
\title{surface validation function}
\usage{
surface_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
surface validation function
}
