% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_validateFunc_headers.R
\name{tcpip_validateFunction}
\alias{tcpip_validateFunction}
\title{tcpip validation function}
\usage{
tcpip_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
tcpip validation function
}
