% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_validateFunc_headers.R
\name{topography_validateFunction}
\alias{topography_validateFunction}
\title{topography type validation function}
\usage{
topography_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
topography type validation function
}
