% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_utils.R
\name{undot}
\alias{undot}
\title{Remove dot on a variable name}
\usage{
undot(varName)
}
\arguments{
\item{varName}{char - Variable Name}
}
\value{
The variable name without a dot.
}
\description{
Internal function.
This function remove a dot behind the variable name : ".foo" => "foo
}
\examples{
\dontrun{
undot(".foo")
}

}
