% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_validateFunc_headers.R
\name{year_validateFunction}
\alias{year_validateFunction}
\title{year validation function}
\usage{
year_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
year validation function
}
