% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_validateFunc_headers.R
\name{zenith_validateFunction}
\alias{zenith_validateFunction}
\title{zenith validation function}
\usage{
zenith_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
zenith validation function
}
