% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_validateFunc_headers.R
\name{A25_validateFunction}
\alias{A25_validateFunction}
\title{A25 validation function}
\usage{
A25_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
A25 validation function
}
