% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-validateFunctions.R
\name{A38_validateFunction}
\alias{A38_validateFunction}
\title{A38 validation function}
\usage{
A38_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
A38 validation function
}
