% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_validateFunc_headers.R
\name{A40_validateFunction}
\alias{A40_validateFunction}
\title{A40 validation function}
\usage{
A40_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
A40 validation function
}
