% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_validateFunc_headers.R
\name{A5_validateFunction}
\alias{A5_validateFunction}
\title{A5 validation function}
\usage{
A5_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
A5 validation function
}
