% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_validateFunc_headers.R
\name{body_validateFunction}
\alias{body_validateFunction}
\title{pyrgeometer body validation function}
\usage{
body_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
pyrgeometer body validation function
}
