% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-utils.R
\name{genericInitialize}
\alias{genericInitialize}
\title{Generic function to create initialize}
\usage{
genericInitialize(vars)
}
\arguments{
\item{vars}{char args needed for the initialize (R notation)}
}
\value{
Initialize function.
}
\description{
Internal function.
Create an initialize R6 function (see R6 documentation).
}
