% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utils.R
\name{genericInitialize}
\alias{genericInitialize}
\title{Generic function to create initialize}
\usage{
genericInitialize(...)
}
\arguments{
\item{...}{list of function args needed for the initialize}
}
\value{
Initialize function.
}
\description{
Internal function.
Create an initialize R6 function (see R6 documentation).
}
