% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_validateFunc_headers.R
\name{I2_validateFunction}
\alias{I2_validateFunction}
\title{I2 validation function}
\usage{
I2_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
I2 validation function
}
