% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-validateFunctions.R
\name{longitude_validateFunction}
\alias{longitude_validateFunction}
\title{longitude validation function}
\usage{
longitude_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
longitude validation function
}
