% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_R6Class_headers.R
\name{LR0001}
\alias{LR0001}
\title{R6 Class representing BSRN LR0001 (Logical Record 0001)}
\arguments{
\item{stationNumber}{Station identification number}

\item{month}{Month of measurement}

\item{year}{Year of measurement}

\item{version}{Version of data}
}
\description{
A LR0001 has a station number, month, year and version.
\itemize{
  \item LR0001$new() : Create a new LR0001 object
  \item isMandatory() : Is variable version is mandatory ?
  \item isMissing() : Is variable version is missing ?
  \item isValuesMissing() : Is any value is missing ?
  \item mandatories() : List of mandatory variable.
  \item missings() : List of missing variables.
  \item setDefault() : Set a variable to default value.
  \item getFormatValue () : Get value with the correct format.
  \item getBsrnFormat() : Get the BSRN format of the object.
  \item showBsrnFormat() : Print the BSRN format.
  \item print() : Generic print method.
}
}
\examples{
lr0001 <- LR0001$new(month = 12, year = 2019, version = 3) # New object.

lr0001$stationNumber = 2 # Affect value.
lr0001$stationNumber
lr0001$params$stationNumber$mandatory # but consult is OK.
lr0001$mandatories() # List of mandatories variables.
lr0001$missings() # List of empty mandatory variables.
lr0001$isValuesMissing() #  There is some missing ?
lr0001$setDefault("month") # Set the default value.
lr0001$month # Missing = NULL
lr0001$month = 1

lr0001$getBsrnFormat() # Get BSRN Format.
lr0001$showBsrnFormat() # Show BSRN Format.
lr0001 # Print the object - a.k.a. print infos.
}
