% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_R6Class_headers.R
\name{LR0001}
\alias{LR0001}
\title{R6 Class representing BSRN LR0001 (Logical Record 0001)}
\arguments{
\item{stationNumber}{Station identification number}

\item{month}{Month of measurement}

\item{year}{Year of measurement}

\item{version}{Version of data}
}
\description{
A LR0001 has a station number, month, year and version.
\itemize{
  \item LR0001$new() : Create a new LR0001 object
  \item isMandatory() : Is variable version is mandatory ?
  \item isMissing() : Is variable version is missing ?
  \item isValuesMissing() : Is any value is missing ?
  \item mandatories() : List of mandatory variable.
  \item missings() : List of missing variables.
  \item setDefault() : Set a variable to default value.
  \item getMissingCode() : Get the missing code for the variable.
  \item getLabel() : Get the label for the variable.
  \item getFormatValue () : Get value with the correct format.
  \item getBsrnFormat() : Get the BSRN format of the object.
  \item showBsrnFormat() : Print the BSRN format.
  \item print() : Generic print method.
}
}
\examples{
lr0001 = LR0001$new(month = 12, year = 2019, version = 3)
lr0001$stationNumber = 2
lr0001$stationNumber
lr0001$mandatories()
lr0001$missings()
lr0001$isValuesMissing()
lr0001$setDefault("month")
lr0001$month
lr0001$month = 1
lr0001$getMissingCode("month")
lr0001$getLabel("month")
lr0001$getBsrnFormat()
lr0001$showBsrnFormat()
lr0001

}
