% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_R6Class_headers.R
\name{LR0002}
\alias{LR0002}
\title{R6 Class representing BSRN LR0002 (Logical Record 0002)}
\arguments{
\item{scientistChange}{Is scientist changed this month ?}

\item{scientistChangeDay}{Day}

\item{scientistChangeHour}{Hour}

\item{scientistChangeMinute}{Minute}

\item{scientistName}{Scientist name *}

\item{scientistTel}{Scientist telephone no. *}

\item{scientistFax}{Scientist fax no. *}

\item{scientistTcpip}{Scientist TCP/IP no.}

\item{scientistMail}{Scientist e-mail address}

\item{scientistAddress}{Scientist address}

\item{deputyChange}{Is deputy changed this month ?}

\item{deputyChangeDay}{Day}

\item{deputyChangeHour}{Hour}

\item{deputyChangeMinute}{Minute}

\item{deputyName}{Deputy name}

\item{deputyTel}{Deputy telephone no.}

\item{deputyFax}{Deputy fax no.}

\item{deputyTcpip}{Deputy TCP/IP no.}

\item{deputyMail}{Deputy e-mail address}

\item{deputyAddress}{Deputy address}
}
\description{
A LR0002 has deputy and scientist informations.
\itemize{
  \item LR0002$new() : Create a new LR0002 object
  \item isMandatory() : Is variable version is mandatory ?
  \item isMissing() : Is variable version is missing ?
  \item isValuesMissing() : Is any value is missing ?
  \item mandatories() : List of mandatory variable.
  \item missings() : List of missing variables.
  \item setDefault() : Set a variable to default value.
  \item getFormatValue () : Get value with the correct format.
  \item getBsrnFormat() : Get the BSRN format of the object.
  \item showBsrnFormat() : Print the BSRN format.
  \item print() : Generic print method.
}
}
\examples{
lr0002 <- LR0002$new(
  scientistName = "John Doe",
  scientistTel = "+1 800 123 4567",
  scientistFax = "+1 800 123 4567",
  scientistMail = "john.doe@example.com",
  scientistAddress = "123 Main Street, Anytown, USA 12345",
  deputyName = "Jane Doe",
  deputyTel = "+1 800 123 4567",
  deputyFax = "+1 800 123 4567",
  deputyMail = "jane.doe@example.com",
  deputyAddress = "123 Main Street, Anytown, USA 12345",
)

lr0002$getBsrnFormat()
lr0002$showBsrnFormat()
lr0002
}
