% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-LR0003.R
\name{LR0003}
\alias{LR0003}
\title{R6 Class representing BSRN LR0003 (Logical Record 0003)}
\arguments{
\item{message}{Messages not to be inserted in the BSRN database}
}
\description{
A LR0003 contain a message.
\itemize{
  \item LR0003$new() : Create a new LR0003 object
  \item isMandatory() : Is variable version is mandatory ?
  \item isMissing() : Is variable version is missing ?
  \item isValuesMissing() : Is any value is missing ?
  \item mandatories() : List of mandatory variable.
  \item missings() : List of missing variables.
  \item setDefault() : Set a variable to default value.
  \item getMissingCode() : Get the missing code for the variable.
  \item getLabel() : Get the label for the variable.
  \item getFormatValue () : Get value with the correct format.
  \item getBsrnFormat() : Get the BSRN format of the object.
  \item showBsrnFormat() : Print the BSRN format.
  \item print() : Generic print method.
}
}
\examples{
lr0003 = LR0003$new(message = "Something important")
lr0003$message = "Some information"
lr0003$message
lr0003$mandatories()
lr0003$missings()
lr0003$isValuesMissing()
lr0003$setDefault("month")
lr0003$month
lr0003$month = 1
lr0003$getMissingCode("month")
lr0003$getLabel("month")
lr0003$getBsrnFormat()
lr0003$showBsrnFormat()
lr0003

}
