% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_bsrnFormats_headers.R
\name{lr0003GetBsrnFormat}
\alias{lr0003GetBsrnFormat}
\title{LR0003 get BSRN format function}
\usage{
lr0003GetBsrnFormat(...)
}
\arguments{
\item{...}{Few more text to put in comment (ex : the LR4000CONST metadata information)}
}
\value{
A char with the BSRN format
}
\description{
LR0003 get BSRN format function
}
