% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_R6Class_headers.R
\name{LR0007}
\alias{LR0007}
\title{R6 Class representing BSRN LR0007 (Logical Record 0007)}
\description{
A LR0007 contain object.
}
\examples{
lr0007 <- LR0007$new()

lr0007$getBsrnFormat()
lr0007$showBsrnFormat()
lr0007$showBsrnFormat(synop = "AAAAA")
lr0007
}
