% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_bsrnFormats_headers.R
\name{lr0007GetBsrnFormat}
\alias{lr0007GetBsrnFormat}
\title{LR0007 get BSRN format function}
\usage{
lr0007GetBsrnFormat(synop = NULL)
}
\arguments{
\item{synop}{the value of the synop (LR0007)}
}
\value{
A char with the BSRN format
}
\description{
LR0007 get BSRN format function
}
