% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-bsrnFormats.R
\name{lr0008GetBsrnFormat}
\alias{lr0008GetBsrnFormat}
\title{LR0008 get BSRN format function}
\usage{
lr0008GetBsrnFormat(change = FALSE, printLr = FALSE)
}
\value{
A char with the BSRN format
}
\description{
LR0008 get BSRN format function
}
