% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_R6Class_datas.R
\name{LR0100}
\alias{LR0100}
\title{R6 Class representing BSRN LR0100 (Logical Record 0100)}
\arguments{
\item{yearMonth}{[mandatory] A string (`'YYYY-MM'`) representing the year and month of measurement.}

\item{global2_avg}{[optional] A numeric vector for global solar radiation averages (W/m^2).}

\item{global2_std}{[optional] A numeric vector for the standard deviation of global solar radiation (W/m^2).}

\item{global2_min}{[optional] A numeric vector for the minimum global solar radiation (W/m^2).}

\item{global2_max}{[optional] A numeric vector for the maximum global solar radiation (W/m^2).}

\item{direct_avg}{[optional] A numeric vector for direct solar radiation averages (W/m^2).}

\item{direct_std}{[optional] A numeric vector for the standard deviation of direct solar radiation (W/m^2).}

\item{direct_min}{[optional] A numeric vector for the minimum direct solar radiation (W/m^2).}

\item{direct_max}{[optional] A numeric vector for the maximum direct solar radiation (W/m^2).}

\item{diffuse_avg}{[optional] A numeric vector for diffuse solar radiation averages (W/m^2).}

\item{diffuse_std}{[optional] A numeric vector for the standard deviation of diffuse solar radiation (W/m^2).}

\item{diffuse_min}{[optional] A numeric vector for the minimum diffuse solar radiation (W/m^2).}

\item{diffuse_max}{[optional] A numeric vector for the maximum diffuse solar radiation (W/m^2).}

\item{downward_avg}{[optional] A numeric vector for downward long-wave radiation averages (W/m^2).}

\item{downward_std}{[optional] A numeric vector for the standard deviation of downward long-wave radiation (W/m^2).}

\item{downward_min}{[optional] A numeric vector for the minimum downward long-wave radiation (W/m^2).}

\item{downward_max}{[optional] A numeric vector for the maximum downward long-wave radiation (W/m^2).}

\item{temperature}{[optional] A numeric vector for air temperature at the instrument height (°C).}

\item{humidity}{[optional] A numeric vector for relative humidity at the instrument height (\%).}

\item{pressure}{[optional] A numeric vector for atmospheric pressure at the instrument height (hPa).}
}
\description{
A LR0100 contain object.
\itemize{
  \item LR0100$new() : Create a new LR0100 object
  \item isMandatory() : Is variable version is mandatory ?
  \item isMissing() : Is variable version is missing ?
  \item isValuesMissing() : Is any value is missing ?
  \item mandatories() : List of mandatory variable.
  \item missings() : List of missing variables.
  \item setDefault() : Set a variable to default value.
  \item getFormatValue () : Get value with the correct format.
  \item getBsrnFormat() : Get the BSRN format of the object.
  \item showBsrnFormat() : Print the BSRN format.
  \item print() : Generic print method.
}
}
\examples{
lr0100 <- LR0100$new()

lr0100$yearMonth <- "2012-01"
lr0100$global2_avg <- rep(1, 44640)

lr0100
lr0100$showBsrnFormat(changed = TRUE)
}
