% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_bsrnFormats_datas.R
\name{lr0100GetBsrnFormat}
\alias{lr0100GetBsrnFormat}
\title{LR0100 get BSRN format function}
\usage{
lr0100GetBsrnFormat(changed)
}
\value{
A char with the BSRN format
}
\description{
LR0100 get BSRN format function
}
