% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_R6Class_datas.R
\name{LR4000}
\alias{LR4000}
\title{R6 Class representing BSRN LR4000 (Logical Record 4000)}
\description{
R6 Class representing BSRN LR4000 (Logical Record 4000)
}
\examples{
lr4000 <- LR4000$new()

lr4000$yearMonth <- "2012-01"
lr4000$bodyT_down <- rep(1, 44640)

lr4000
lr4000$showBsrnFormat(changed = TRUE)
}
