% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_bsrnFormats_headers.R
\name{lr4000constGetBsrnFormat}
\alias{lr4000constGetBsrnFormat}
\title{LR4000CONST get BSRN format function}
\usage{
lr4000constGetBsrnFormat(method = 1)
}
\value{
A char with the BSRN format
}
\description{
LR4000CONST get BSRN format function
}
