% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-validateFunctions.R
\name{month_validateFunction}
\alias{month_validateFunction}
\title{month validation function}
\usage{
month_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
month validation function
}
