% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_validateFunc_headers.R
\name{telephone_validateFunction}
\alias{telephone_validateFunction}
\title{telephone validation function}
\usage{
telephone_validateFunction(value)
}
\arguments{
\item{value}{value to validate}
}
\description{
telephone validation function
}
